Initial backend project

+++ src/main/java/com/shop/tailors/repository/CustomerRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.shop.tailors.entity.Customer;
 
 @Repository
-public interface CustomerRepository  extends JpaRepository<Customer, Integer>{
+public interface CustomerRepository  extends JpaRepository<Customer, Long>{
